Change structure of entity package to break import cycle
